Move the core implementation to xsd.core paackage
