+++ xsd-core/src/main/java/module-info.java
@@ -3,7 +3,7 @@ module io.ballerina.xsdtorecordconverter {
     requires io.ballerina.parser;
     requires io.ballerina.formatter.core;
 
-    exports io.ballerina.xsdtorecordconverter;
-    exports io.ballerina.xsdtorecordconverter.visitor;
-    exports io.ballerina.xsdtorecordconverter.component;
+    exports io.ballerina.xsd.core;
+    exports io.ballerina.xsd.core.visitor;
+    exports io.ballerina.xsd.core.component;
 }
