+++ xsd-core/src/test/java/io/ballerina/xsd/core/XSDToRecordTest.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package io.ballerina.xsdtorecordconverter;
+package io.ballerina.xsd.core;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
